REFACTOR CODE, AUTOPEP8 AND CLUSTER HPC CODEBASE
Refactored code to include a module for subclassing experiment types
that we would run with the TVB simulations.

Ran autopep8 on the tvbsim dir.

Included the cluster hpc codebase that has the clinez and pz regions
for each tng patient and can bbe run for exp008,009 and 010, where
we have just original connectivities and varying the clinez/pz,
across patient shuffled connectivity and then within patient
preserving weights shuffling.

To Do:
1. Continue testing and improving robustness of the simulation code.
2. Possibly come up with a better algo. for peak detection used
for seizure detection on the z time series.